+++ core-bundle/src/Resources/contao/config/constants.php
@@ -10,7 +10,7 @@
 
 // Core version
 \define('VERSION', '4.4');
-\define('BUILD', '16');
+\define('BUILD', '17');
 \define('LONG_TERM_SUPPORT', true);
 
 // Link constants